+++ packages/yourloops/components/dialogs/join-team/privacy-policy-confirm.tsx
@@ -53,11 +53,12 @@ export const PrivacyPolicyConfirm = (props: ConfirmTeamProps): JSX.Element => {
   const { onCompleteStep, team, onClickCancel, inProgress } = props
   const [policyChecked, setPolicyChecked] = useState<boolean>(false)
   const theme = useTheme()
+  const privacyPolicy = t('privacy-policy')
 
   const handleChangeChecked = (event: React.ChangeEvent<HTMLInputElement>): void => {
     setPolicyChecked(event.target.checked)
   }
-  const privacyPolicy = t('privacy-policy')
+
   const linkPrivacyPolicy = (
     <Link
       aria-label={privacyPolicy}