+++ packages/yourloops/test/integration/mock/medical-files.api.mock.ts
@@ -27,35 +27,35 @@
 
 import MedicalFilesApi from '../../../lib/medical-files/medical-files.api'
 import { monitoredPatientId } from './patient.api.mock'
+import { teamTwo } from './team.api.mock'
 
 export const mockMedicalFilesAPI = () => {
-  jest.spyOn(MedicalFilesApi, 'getPrescriptions').mockResolvedValue([{
-    id: 'prescriptionMocked',
-    name: 'patientPrescription',
-    patientId: monitoredPatientId,
-    teamId: 'team1234',
-    prescriptorId: 'prescriptorId',
-    link: 'https://linkToPrescription',
-    uploadedAt: '2022-01-02T08:34:06.898Z'
-  }])
   jest.spyOn(MedicalFilesApi, 'getMedicalReports').mockResolvedValue([{
     id: 'medicalReportId',
     authorId: 'authorId',
-    creationDate: '2022-01-02T08:34:06.898Z',
+    creationDate: '2022-01-10T08:34:06.898Z',
     patientId: monitoredPatientId,
-    teamId: 'team1234',
+    teamId: teamTwo.id,
+    teamName: teamTwo.name,
     diagnosis: 'whatever diagnosis',
     progressionProposal: 'whatever proposal',
-    trainingSubject: 'here is the subject'
+    trainingSubject: 'here is the subject',
+    number: 1,
+    authorFirstName: 'Vishnou',
+    authorLastName: 'Lapaix'
   },
   {
     id: 'medicalReportId2',
     authorId: 'authorId',
     creationDate: '2022-01-02T10:30:00.000Z',
     patientId: monitoredPatientId,
-    teamId: 'team1234',
+    teamId: teamTwo.id,
+    teamName: teamTwo.name,
     diagnosis: 'whatever diagnosis 2 ',
     progressionProposal: 'whatever proposal 2',
-    trainingSubject: 'here is the subject 2'
+    trainingSubject: 'here is the subject 2',
+    number: 2,
+    authorFirstName: 'Vishnou',
+    authorLastName: 'Lapaix'
   }])
 }